requirements.txt
pytest==5.3.2
pytest-cov==2.8.1
pylint==2.4.4
flake8==3.7.9
Flask==1.1.1
bandit==1.6.2
Flask-SQLAlchemy==2.4.1
mysql-connector-python==8.0.19
Flask-APScheduler==1.11.0
flask-cors==3.0.8
Sphinx==2.4.1
sphinx-rtd-theme==0.4.3
fpdf==1.7.2
